Move find_closest_sgRNA_cut_point function from CORE into data_prep
